+++ filetype/file_type.go
@@ -1,0 +1,9 @@
+package filetype
+
+const (
+	PEM FileType = iota
+	DER
+	ENG
+)
+
+type FileType int